+++ src/main/java/com/github/games647/flexiblelogin/FlexibleLogin.java
@@ -52,6 +52,7 @@ import org.spongepowered.api.config.DefaultConfig;
 import org.spongepowered.api.event.Listener;
 import org.spongepowered.api.event.game.state.GameInitializationEvent;
 import org.spongepowered.api.event.game.state.GamePreInitializationEvent;
+import org.spongepowered.api.event.game.state.GameStoppedServerEvent;
 import org.spongepowered.api.plugin.Plugin;
 import org.spongepowered.api.plugin.PluginContainer;
 import org.spongepowered.api.text.Text;
@@ -157,6 +158,12 @@ public class FlexibleLogin {
         //register events
         game.getEventManager().registerListeners(this, new PlayerConnectionListener());
         game.getEventManager().registerListeners(this, new PreventListener());
+    }
+
+    @Listener
+    public void onDisable(GameStoppedServerEvent gameStoppedEvent) {
+        //run this task sync in order let it finish before the process ends
+        database.close();
     }
 
     public Settings getConfigManager() {